Refactor preferences -> move them to subcontroller, subroute
