+++ app/assets/javascripts/users/settings/account/preferences.js
